Rewrite entries into a loop to remove recursion
Use an explicit stack to replace application stack for recursion.
This is slightly faster, allows for manual TCO but requires an extra allocation
for the stack. However this can be optimized away for the common case with doing
a manual preallocation on the application stack. However this needs to be of a
fixed size. 20 was used since at size 2^20 allocation of the return slice seems
to dominate.